Refactored the database and Models
Splited the Database and Models in two separated files
