+++ python/flask/code/app/database.py
@@ -1,0 +1,6 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+from . import app
+
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
